Put Patreon NotificationHandlers into Feature folder.
