+++ backend/Veloci.Logic/Bot/Discord/DiscordMessageEventHandler.cs
@@ -3,7 +3,6 @@ using MediatR;
 using Serilog;
 using Veloci.Data.Domain;
 using Veloci.Data.Repositories;
-using Veloci.Logic.Features.Patreon.Notifications;
 using Veloci.Logic.Helpers;
 using Veloci.Logic.Notifications;
 using Veloci.Logic.Services;
@@ -22,9 +21,7 @@ public class DiscordMessageEventHandler :
     INotificationHandler<YearResults>,
     INotificationHandler<DayStreakAchievements>,
     INotificationHandler<DayStreakPotentialLose>,
-    INotificationHandler<GotAchievements>,
-    INotificationHandler<NewPatreonSupporterNotification>,
-    INotificationHandler<MonthlyPatreonSupportersNotification>
+    INotificationHandler<GotAchievements>
 {
     private static readonly ILogger _log = Log.ForContext<DiscordMessageEventHandler>();
 
@@ -183,44 +180,6 @@ public class DiscordMessageEventHandler :
     public async Task Handle(GotAchievements notification, CancellationToken cancellationToken)
     {
         var message = _messageComposer.AchievementList(notification.Results);
-        await _discordBot.SendMessageAsync(message);
-    }
-
-    public async Task Handle(NewPatreonSupporterNotification notification, CancellationToken cancellationToken)
-    {
-        var message = $"🎉 Вітаємо нового підтримувача Patreon: **{notification.Supporter.Name}**";
-        if (!string.IsNullOrEmpty(notification.Supporter.TierName))
-        {
-            message += $" ({notification.Supporter.TierName})";
-        }
-        message += "! Дякуємо за підтримку! ❤️";
-        
-        await _discordBot.SendMessageAsync(message);
-    }
-
-    public async Task Handle(MonthlyPatreonSupportersNotification notification, CancellationToken cancellationToken)
-    {
-        if (!notification.Supporters.Any())
-            return;
-
-        var message = $"📊 **Щомісячні підтримувачі Patreon** ({notification.Supporters.Count}):\n\n";
-        
-        var groupedByTier = notification.Supporters
-            .GroupBy(s => s.TierName ?? "Невідомий рівень")
-            .OrderByDescending(g => g.Average(s => s.Amount ?? 0));
-
-        foreach (var tierGroup in groupedByTier)
-        {
-            message += $"**{tierGroup.Key}:**\n";
-            foreach (var supporter in tierGroup.OrderBy(s => s.Name))
-            {
-                message += $"• {supporter.Name}\n";
-            }
-            message += "\n";
-        }
-
-        message += "Дякуємо всім за постійну підтримку! 🙏";
-        
         await _discordBot.SendMessageAsync(message);
     }
 }
