+++ backend/Veloci.Logic/Bot/Telegram/TelegramMessageEventHandler.cs
@@ -1,6 +1,5 @@
 ﻿using Hangfire;
 using MediatR;
-using Veloci.Logic.Features.Patreon.Notifications;
 using Veloci.Logic.Helpers;
 using Veloci.Logic.Notifications;
 using Veloci.Logic.Services;
@@ -19,9 +18,7 @@ public class TelegramMessageEventHandler :
     INotificationHandler<YearResults>,
     INotificationHandler<DayStreakAchievements>,
     INotificationHandler<DayStreakPotentialLose>,
-    INotificationHandler<GotAchievements>,
-    INotificationHandler<NewPatreonSupporterNotification>,
-    INotificationHandler<MonthlyPatreonSupportersNotification>
+    INotificationHandler<GotAchievements>
 {
     private readonly TelegramMessageComposer _messageComposer;
 
@@ -131,44 +128,6 @@ public class TelegramMessageEventHandler :
     public async Task Handle(GotAchievements notification, CancellationToken cancellationToken)
     {
         var message = _messageComposer.AchievementList(notification.Results);
-        await TelegramBot.SendMessageAsync(message);
-    }
-
-    public async Task Handle(NewPatreonSupporterNotification notification, CancellationToken cancellationToken)
-    {
-        var message = $"🎉 Вітаємо нового підтримувача Patreon: *{notification.Supporter.Name}*";
-        if (!string.IsNullOrEmpty(notification.Supporter.TierName))
-        {
-            message += $" ({notification.Supporter.TierName})";
-        }
-        message += "! Дякуємо за підтримку! ❤️";
-        
-        await TelegramBot.SendMessageAsync(message);
-    }
-
-    public async Task Handle(MonthlyPatreonSupportersNotification notification, CancellationToken cancellationToken)
-    {
-        if (!notification.Supporters.Any())
-            return;
-
-        var message = $"📊 *Щомісячні підтримувачі Patreon* ({notification.Supporters.Count}):\n\n";
-        
-        var groupedByTier = notification.Supporters
-            .GroupBy(s => s.TierName ?? "Невідомий рівень")
-            .OrderByDescending(g => g.Average(s => s.Amount ?? 0));
-
-        foreach (var tierGroup in groupedByTier)
-        {
-            message += $"*{tierGroup.Key}:*\n";
-            foreach (var supporter in tierGroup.OrderBy(s => s.Name))
-            {
-                message += $"• {supporter.Name}\n";
-            }
-            message += "\n";
-        }
-
-        message += "Дякуємо всім за постійну підтримку! 🙏";
-        
         await TelegramBot.SendMessageAsync(message);
     }
 }
